Cleanup Property impl systems
